Update `git neco review` command
`git neco review` command can change `draft` branch to `ready for review` branch if `draft` branch exists
